restructure files and imports and make daemon executable
note this removes some unused imports
